Added auto level change (issues!)
The game changes the level automatically once the objects are cleared.
Should think about bigger scene!

(NOTE)Lifes label is used currently to display the objects on the scene
for easier troubleshooting.

(ISSUE) Seems that sometimes there are asteroids OUTSIDE of the field!

(ISSUE) For some reason scene-> clear(); crashes the game when changing
the level but not on newgame. For this reason currently new ship will
spawn on change level but the old one will stay and will be the one to
be moved around.